refactor(cli): extract doctor diagnostics boundary
